+++ doc.go
@@ -1,0 +1,3 @@
+//go:generate go run -tags gendocs ./cmd/certkit
+
+package certkit